commit f4a458619e57465e9622848a47d014ee65cfb076
    17515 ucodeadm could handle binary AMD microcode files
commit a86602b15b9c56eb076e48653db3cf239babfec8
    17495 NFSv4.2 unsupported operations missing reply encoding
commit d852f8ff587f58ad4c891927373169c1cec33ecf
    16338 pynfs/nfs4.1: COUR2 st_courtesy.testLockSleepLock FAILURE
    16346 pynfs/nfs4.1: SEQ6 st_sequence.testRequestTooBig FAILURE
    16355 pynfs/nfs4.1: SEQ10b st_sequence.testReplayCache007 FAILURE
    16356 pynfs/nfs4.1: SEQ9e st_sequence.testReplayCache005 FAILURE
    16357 pynfs/nfs4.1: CSESS28 and CSESS29 pynfs tests FAILURE
    16358 pynfs/nfs4.1: RECC3 st_reclaim_complete.testOpenBeforeRECC FAILURE
    16359 pynfs/nfs4.1: refactor to ONE_STATEID and ZERO_STATEID
    16360 pynfs/nfs4.1: OP_CLOSE should return special invalid stateid
    16361 pynfs/nfs4.1: RFS4 doesn't need RPC_ALL
    16362 pynfs/nfs4.1: Add support for CURRENT_STATEID
    16363 pynfs/nfs4.1: Add suppport for FREE_STATEID
    16364 pynfs/nfs4.1: COUR3 COUR5 COUR6 FAILURE
commit 2fba95bdf3072aa76182a28a34bc0688de2676ba
    17519 nvme panics on device with no namespaces

+++ usr/src/common/ucode/ucode_utils_intel.h
@@ -10,7 +10,7 @@
  */
 
 /*
- * Copyright 2023 Oxide Computer Company
+ * Copyright 2025 Oxide Computer Company
  */
 
 #ifndef _COMMON_UCODE_UTILS_INTEL_H
@@ -26,7 +26,7 @@ extern ucode_errno_t ucode_header_validate_intel(ucode_header_intel_t *);
 extern uint32_t ucode_checksum_intel(uint32_t, uint32_t, uint8_t *);
 extern uint32_t ucode_checksum_intel_extsig(ucode_header_intel_t *,
     ucode_ext_sig_intel_t *);
-extern ucode_errno_t ucode_validate_intel(uint8_t *, int);
+extern ucode_errno_t ucode_validate_intel(uint8_t *, size_t);
 
 #ifdef __cplusplus
 }
